Start refactor by breaking apart the main index file
