+++ index.js
@@ -2,10 +2,8 @@
 //
 // SPDX-License-Identifier: MIT
 
-const fs = require('fs')
 const url = require('url')
 const path = require('path')
-const assert = require('assert')
 
 const express = require('express')
 const app = express()
@@ -20,19 +18,27 @@ const Logger = require('./lib/Logger')
 
 const template = require('./app/template')
 
-const utils = require('./lib/utils')
-const paths = require('./lib/paths')
 const network = require('./lib/network')
-const electron = require('./lib/electron')
 const platform = require('./lib/platform')
 const apiRoutes = require('./lib/routes')
 
 const pkg = require('./package.json')
+const config = require('./lib/config')
+
+/*
+Do required initialization
+*/
+require('./lib/init-common')
+if (platform.isElectron()) {
+  require('./lib/init-electron')
+} else {
+  require('./lib/init-node')
+}
 
 /**
  * @type { Number }
  */
-const DEFAULT_HTTP_PORT = 5544
+const DEFAULT_HTTP_PORT = config.defaults.HTTP_PORT
 const ASSETS = require('./assets.json')
 
 /**
@@ -43,70 +49,9 @@ const ASSETS = require('./assets.json')
 */
 const WORKSPACE_TEARDOWN_MIN_THRESHOLD_MS = 20000
 
-/**
- * Verify that an assets file is
- * created before running the app,
- * hashes are used in order to eliminate
- * caching issues
- */
-;(function () {
-  const assetsExist = fs.existsSync(path.join(__dirname, './assets.json'))
-  assert(
-    assetsExist,
-    'No assets file found, the project must be built before it\'s run: \'npm build\''
-  )
-})()
-
-/**
- * Create the plugin directories
- * if they don't already exist
- */
-;(function () {
-  Logger.debug('Creating plugin directory')
-  utils.createDirectoryRecursively(paths.plugins)
-})()
-
-/**
- * Remove and recreate the temporary directory
- * in order to make sure that it's cleared and
- * exists
- */
-;(function () {
-  Logger.debug('Recreating temporary directory')
-  try {
-    fs.rmSync(paths.temp, { force: true, recursive: true })
-  } catch (err) {
-    Logger.warn('Failed to remove temporary files directory', err)
-  }
-  utils.createDirectoryRecursively(paths.temp)
-})()
-
-/**
- * Restore user defaults into
- * the user defaults-state
- */
-;(async function () {
-  Logger.debug('Restoring user deafults', paths.userDefaults)
-  let json
-  try {
-    const data = fs.readFileSync(paths.userDefaults, { encoding: 'utf8' })
-    json = JSON.parse(data || '{}')
-
-    UserDefaults.apply({
-      ...json
-    })
-  } catch (err) {
-    Logger.warn('Failed to restore user defaults, maybe it\'s the first time the application is running', err)
-  } finally {
-    UserDefaults.apply({
-      httpPort: process.env.PORT || json?.httpPort || DEFAULT_HTTP_PORT
-    })
-  }
-})()
-
 /*
 These constants depend on the UserDefaults-state and
-MUST be declared AFTER the state is restored
+MUST be declared AFTER initialization
 */
 const HTTP_PORT = UserDefaults.data.httpPort || DEFAULT_HTTP_PORT
 const HTTP_BIND_ADDR = UserDefaults.data.httpBindToAll ? '0.0.0.0' : 'localhost'
@@ -285,49 +230,3 @@ app.use((err, req, res, next) => {
       description: _err.message
     })
 })
-
-/*
-Setup a new window if running
-in an electron context
-*/
-;(async function () {
-  if (!platform.isElectron()) {
-    return
-  }
-  await electron.isReady()
-
-  if (electron.wasOpenedByFile()) {
-    return
-  }
-
-  electron.initWindow(`http://localhost:${UserDefaults.data.httpPort}`)
-})()
-
-/*
-Write the user defaults-state to disk
-before the process exits
-*/
-;(function () {
-  function writeUserDeafults () {
-    Logger.debug('Writing user defaults to disk')
-    fs.writeFileSync(paths.userDefaults, JSON.stringify(UserDefaults.data))
-  }
-
-  if (platform.isElectron()) {
-    electron.app.once('will-quit', () => {
-      writeUserDeafults()
-    })
-  } else {
-    process.on('exit', () => writeUserDeafults())
-
-    process.on('SIGTERM', () => {
-      writeUserDeafults()
-      process.exit(0)
-    })
-
-    process.on('SIGINT', () => {
-      writeUserDeafults()
-      process.exit(0)
-    })
-  }
-})()
