+++ lib/config.js
@@ -1,0 +1,7 @@
+// SPDX-FileCopyrightText: 2024 Sveriges Television AB
+//
+// SPDX-License-Identifier: MIT
+
+exports.defaults = {
+  HTTP_PORT: 5544
+}
